% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoms.R
\name{geom_borrow}
\alias{geom_borrow}
\title{Plot a Bayesian basket trial's posterior distribution after borrowing}
\usage{
geom_borrow(..., design, n, r)
}
\arguments{
\item{...}{Further arguments to be passed to `geom_function`.}

\item{design}{An object created with one of the \code{setup} functions.}

\item{n}{The sample size per basket.}

\item{r}{Vector of responses.}
}
\value{
A list of ggplot layers of type `geom_function`.
}
\description{
Plot a Bayesian basket trial's posterior distribution after borrowing
}
\examples{
# Example for a basket trial with Fujikawa's Design
design <- setup_fujikawa(k = 3, p0 = 0.2)
n <- 20
r <- c(4, 5, 2)
lambda <- 0.99
epsilon <- 2
tau <- 0.5
# One facet per basket
ggplot() +
    geom_borrow(design, n, r, lambda, epsilon, tau) +
    facet_wrap(vars(basket))
# Colour different baskets
ggplot() +
    geom_borrow(aes(colour = basket), design, n, r, lambda, epsilon, tau)
}
