% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_details.R
\name{get_details.ebcomb}
\alias{get_details.ebcomb}
\title{Get Details of a Basket Trial Simulation with the ebcomb Design}
\usage{
\method{get_details}{ebcomb}(design, n, p1, lambda, iter = 1000, data = NULL, ...)
}
\arguments{
\item{design}{An object of class \code{ebcomb}.}

\item{n}{The sample size per basket.}

\item{p1}{Probabilities under the alternative hypothesis.}

\item{lambda}{The posterior probability threshold.}

\item{iter}{The number of iterations in the simulation.}

\item{data}{A data matrix with k column with the number of responses for each
basket. If \code{data} is used, then \code{iter} and \code{p1} are ignored.}

\item{...}{Further arguments.}
}
\value{
A list containing the rejection probabilites, posterior means,
mean squared errors and mean limits of HDI intervals for all baskets.
}
\description{
Get Details of a Basket Trial Simulation with the ebcomb Design
}
\examples{
design <- setup_ebcomb(k = 3, p0 = 0.2)
get_details(design = design, n = 20, p1 = 0.5, lambda = 0.95, iter = 100)
}
