% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_evaluation.R
\name{get_evaluation.bhm}
\alias{get_evaluation.bhm}
\title{Evaluate a BHM Basket Trial}
\usage{
\method{get_evaluation}{bhm}(
  design,
  n,
  r,
  lambda,
  level = 0.95,
  tau_scale,
  n_mcmc = 10000,
  ...
)
}
\arguments{
\item{design}{An object of class \code{bhm}.}

\item{n}{The sample sizes of the baskets. A vector must be used for varying sample sizes.}

\item{r}{The number of responses observed in each basket.}

\item{lambda}{The posterior probability threshold.}

\item{level}{Level of the credibility intervals.}

\item{tau_scale}{Standard deviation of the half normal prior distribution
for the variance of the thetas.}

\item{n_mcmc}{Number of MCMC samples.}

\item{...}{Further arguments.}

\item{iter}{The number of iterations in the simulation. Is ignored if
\code{data} is specified.}
}
\value{
A list containing the point estimates of the basket-specific response rates.
}
\description{
Evaluate a BHM Basket Trial
}
\examples{
design <- setup_bhm(k = 3, p0 = 0.2, p_target = 0.5)

get_evaluation(design = design, n = c(20, 20, 20), r = c(10, 15, 5),
  lambda = 0.95, tau_scale = 1, iter = 100)

# Unequal sample sizes
get_evaluation(design = design, n = c(15, 20, 25), r = c(10, 15, 17),
  lambda = 0.95, tau_scale = 1, iter = 100)
}
