% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_results.R
\name{get_results.bma}
\alias{get_results.bma}
\title{Get Results for Simulation of a Basket Trial with the BMA Design}
\usage{
\method{get_results}{bma}(design, n, p1, lambda, pmp0, iter = 1000, data = NULL, ...)
}
\arguments{
\item{design}{An object of class \code{bma}.}

\item{n}{The sample size per basket.}

\item{p1}{Probabilities under the alternative hypothesis.}

\item{lambda}{The posterior probability threshold.}

\item{pmp0}{Power parameter that is used to compute prior probabilities. See
\code{\link[bmabasket]{bma}} for details.}

\item{iter}{The number of iterations in the simulation.}

\item{data}{A data matrix with k column with the number of responses for each
basket. If \code{data} is used, then \code{iter} and \code{p1} are ignored.}

\item{...}{Further arguments.}
}
\value{
A matrix of results with \code{iter} rows. A 0 means, that the
null hypothesis that the response probability exceeds \code{p0} was not
rejected, a 1 means, that the null hypothesis was rejected.
}
\description{
Get Results for Simulation of a Basket Trial with the BMA Design
}
\examples{
design <- setup_bma(k = 3, p0 = 0.2)
get_results(design, n = 20, p1 = c(0.2, 0.5, 0.5), lambda = 0.95,
  pmp0 = 1, iter = 100)
}
