% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_results.R
\name{get_results.cppgen}
\alias{get_results.cppgen}
\title{Get Results for Simulation of a Basket Trial with a Generalized Calibrated
Power Prior Design}
\usage{
\method{get_results}{cppgen}(
  design,
  n,
  p1,
  lambda,
  tune_a,
  tune_b,
  epsilon,
  iter = 1000,
  data = NULL,
  ...
)
}
\arguments{
\item{design}{An object of class \code{cppgen}.}

\item{n}{The sample size per basket.}

\item{p1}{Probabilities under the alternative hypothesis.}

\item{lambda}{The posterior probability threshold.}

\item{tune_a}{First tuning parameter that determines the amount of borrowing
based on pairwise similarity between baskets.}

\item{tune_b}{Second tuning parameter that determines the amount of borrowing
based on pairwise similarity between baskets.}

\item{epsilon}{Tuning parameter that determines the amount of borrowing based
on overall heterogeneity.}

\item{iter}{The number of iterations in the simulation.}

\item{data}{A data matrix with k column with the number of responses for each
basket. If \code{data} is used, then \code{iter} and \code{p1} are ignored.}

\item{...}{Further arguments.}
}
\value{
A matrix of results with \code{iter} rows. A 0 means, that the
null hypothesis that the response probability exceeds \code{p0} was not
rejected, a 1 means, that the null hypothesis was rejected.
}
\description{
Get Results for Simulation of a Basket Trial with a Generalized Calibrated
Power Prior Design
}
\examples{
design <- setup_cppgen(k = 3, p0 = 0.2)
get_results(design = design, n = 20, p1 = c(0.2, 0.5, 0.5), lambda = 0.95,
  tune_a = 1, tune_b = 1, epsilon = 2, iter = 100)
}
