% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_design.R
\name{opt_design}
\alias{opt_design}
\title{Optimize a Basket Trial Design}
\usage{
opt_design(
  design,
  n,
  alpha,
  design_params = list(),
  scenarios,
  prec_digits,
  iter = 1000,
  ...
)
}
\arguments{
\item{design}{An object created with one of the \code{setup} functions.}

\item{n}{The sample size per basket.}

\item{alpha}{The one-sided significance level.}

\item{design_params}{A list of params that is specific to the class of
\code{design}.}

\item{scenarios}{A matrix of scenarios.}

\item{prec_digits}{Number of decimal places that are considered when
adjusting lambda.}

\item{iter}{The number of iterations in the simulation.}

\item{...}{Further arguments.}
}
\value{
A matrix with the expected number of correct decisions.
}
\description{
Optimize a Basket Trial Design
}
