% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim.R
\name{optim.basket}
\alias{optim.basket}
\title{Optimize a basket trial design}
\usage{
optim.basket(design, par, method, control, utility, constraint, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Optimize a basket trial design
}
