% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms.R
\name{optimization_optim_sa}
\alias{optimization_optim_sa}
\title{Simulated annealing from the `optimization` package}
\usage{
optimization_optim_sa(
  fun,
  trace = FALSE,
  start,
  maximization = FALSE,
  lower,
  upper,
  control = list()
)
}
\arguments{
\item{fun}{A function that should be optimized. The function should only
accept one parameter `x`.}

\item{trace}{A logical, should the trace of the optimization algorithm be
returned?}

\item{start}{numerical, a vector of start parameters.}

\item{maximization}{logical, should the function be maximized}

\item{lower}{numerical, a vector of lower bounds of the parameters.}

\item{upper}{numerical, a vector of upper bounds of the parameters.}

\item{control}{a list of further arguments to be passed to `optim_sa()`}
}
\value{
a list consisting of the optimization result vector, the optimal
 utility value, and the trace of the optimization algorithm
}
\description{
This function is wrapper of `optimization::optim_sa()`.
}
