% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{setup_binomial}
\alias{setup_binomial}
\title{Set Up Frequentist Binomial Design Object}
\usage{
setup_binomial(k, p0, pool = FALSE)
}
\arguments{
\item{k}{The number of baskets.}

\item{p0}{A common probability under the null hypothesis.}

\item{pool}{A logical, should data from all baskets be pooled? Default is
\code{FALSE}. Currently \code{FALSE} is the only working option, reserved for
future implementation.}
}
\value{
An S3 object of class \code{binomial}
}
\description{
Set Up Frequentist Binomial Design Object
}
\details{
The class \code{binomial} implements a basket trial
design, in which each null hypothesis is tested using the
frequentist binomial test without multiplicity correction. All baskets are
either tested separately (the default) or pooled (not implemented yet).
}
\examples{
design_binomial <- setup_binomial(k = 3, p0 = 0.2)
}
