% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{setup_cppgen}
\alias{setup_cppgen}
\title{Setup Generalized Calibrated Power Prior Design Object}
\usage{
setup_cppgen(k, p0, shape1 = 1, shape2 = 1)
}
\arguments{
\item{k}{The number of baskets.}

\item{p0}{A common probability under the null hypothesis.}

\item{shape1}{First common shape parameter of the beta prior.}

\item{shape2}{Second common shape parameter of the beta prior.}
}
\value{
An S3 object of class \code{cppgen}
}
\description{
Setup Generalized Calibrated Power Prior Design Object
}
\examples{
design_cppgen <- setup_cppgen(k = 3, p0 = 0.2)
}
