% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{setup_ebcomb}
\alias{setup_ebcomb}
\title{Setup ebcomb Design Object}
\usage{
setup_ebcomb(k, p0, shape1 = 1, shape2 = 1)
}
\arguments{
\item{k}{The number of baskets.}

\item{p0}{A common probability under the null hypothesis.}

\item{shape1}{First common shape parameter of the beta prior.}

\item{shape2}{Second common shape parameter of the beta prior.}
}
\value{
An S3 object of class \code{bma}
}
\description{
Creates an object of class \code{ebcomb}.
}
\details{
The class \code{ebcomb} implements an empirical Bayes method
by Gravestock & Held (2019) which was proposed for borrowing strength
from multiple external studies.
}
\examples{
design_bma <- setup_bma(k = 3, p0 = 0.2)
}
\references{
Gravestock, I., & Held, L. (2019). Power priors based on
multiple historical studies for binary outcomes. Biometrical Journal, 61(5),
1201-1218.
}
