% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms.R
\name{stats_optim_sann}
\alias{stats_optim_sann}
\title{Simulated annealing from the `optim` function of the `stats` library}
\usage{
stats_optim_sann(fun, trace = FALSE, start, maximization = FALSE, control)
}
\arguments{
\item{fun}{A function that should be optimized. The function should only
accept one parameter `x`.}

\item{trace}{doesn't do anything}

\item{start}{numerical, a vector of start parameters.}

\item{maximization}{logical, should the function be maximized}

\item{control}{a list of further arguments to be passed to `optim()`}
}
\value{
a list consisting of the optimization result vector, the optimal
 utility value, and the trace of the optimization algorithm
}
\description{
This function is wrapper of `stats::optim(method = "SANN")`.
}
