% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{u_powfwer_discont_bound}
\alias{u_powfwer_discont_bound}
\title{Utility function: Discontinuous power function with type-I error penalty and
boundaries on the parameters}
\usage{
u_powfwer_discont_bound(design, x, detail_params, alpha, lower, upper)
}
\arguments{
\item{design}{An object created with one of the \code{setup} functions.}

\item{x}{A named list of the design parameters to be optimized.}

\item{detail_params}{A named list of parameters that need to be supplied to
`get_details()`.}

\item{lower}{numerical, a vector of lower bounds of the parameters.}

\item{upper}{numerical, a vector of upper bounds of the parameters.}
}
\value{
a numerical, the parameter combination's utility
}
\description{
Utility function: Discontinuous power function with type-I error penalty and
boundaries on the parameters
}
\examples{
design <- setup_fujikawa(k = 3, p0 = 0.2, shape1 = 1, shape2 = 1)
u_powfwer_discont_bound(design,
                  x = list(lambda = 0.99, epsilon = 2, tau = 0.5),
                  detail_params = list(n = 20, p1 = c(0.5, 0.2, 0.2),
                                       logbase = exp(1), exact = TRUE),
                  alpha = 0.05,
                  lower = list(lambda = 0, epsilon = 1, tau = 0),
                  upper = list(lambda = 1, epsilon = 10, tau = 1))
}
