% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basksim-package.R
\docType{package}
\name{basksim-package}
\alias{basksim}
\alias{basksim-package}
\title{basksim: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sabrinaschmitt/basksim}
  \item Report bugs at \url{https://github.com/sabrinaschmitt/basksim/issues}
}

}
\keyword{internal}
