% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basksim-package.R
\docType{package}
\name{basksim-package}
\alias{basksim}
\alias{basksim-package}
\title{basksim: Simulation-Based Evaluation of Basket Trials}
\description{
Provides a unified syntax for the simulation-based comparison of different single-stage basket trial designs with a binary endpoint and baskets of equal or unequal sample sizes. Methods include the designs by Baumann et al. (2024) \doi{10.48550/arXiv.2309.06988}, Fujikawa et al. (2020) \doi{10.1002/bimj.201800404}, Berry et al. (2020) \doi{10.1177/1740774513497539}, Neuenschwander et al. (2016) \doi{10.1002/pst.1730} and Psioda et al. (2021) \doi{10.1093/biostatistics/kxz014}. For the latter three designs, the functions are mostly wrappers for functions provided by the packages 'bhmbasket' and 'bmabasket'. In addition, the package enables the application of the design of Ollier et al. (2020) \doi{10.1177/0962280219886609} in the basket trial setting and provides a new method that combines the two designs of Ollier et al. (2020) and Baumann et al. (2024).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sabrinaschmitt/basksim}
  \item Report bugs at \url{https://github.com/sabrinaschmitt/basksim/issues}
}

}
\keyword{internal}
