% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_details.R
\name{get_details.binomial}
\alias{get_details.binomial}
\title{Get Details of a Basket Trial with the Frequentist Binomial Design}
\usage{
\method{get_details}{binomial}(design, n, p1 = NULL, alpha = 0.025, ...)
}
\arguments{
\item{design}{An object of class \code{binomial}.}

\item{n}{The sample sizes of the baskets. A vector must be used for varying sample sizes.}

\item{p1}{Probabilities used for the simulation. If \code{NULL}
then all probabilities are set to p0.}

\item{alpha}{The one-sided significance level.}

\item{...}{Further arguments.}
}
\value{
A list containing the rejection probabilities, critical values
and expected number of correct decisions. Critical values \eqn{c} are defined
so that the null hypothesis is rejected if the observed number of responses
\eqn{r} is greater than \eqn{c}, i.e. \eqn{r > c} rejects \eqn{H_0}.
The nominal FWER is the (theoretical) FWER of a multiple testing problem
with k hypothesis tests at significance level \eqn{\alpha}. The actual FWER is
usually lower than the nominal FWER, as the binomial test does not exhaust
its significance level.
}
\description{
This basic frequentist design conducts a separate binomial test for each
basket. All tests are one-sided, and the alternative is greater than the
null hypothesis. These details are calculated exactly, not simulated.
}
\examples{
design <- setup_binomial(k = 3, p0 = 0.2)
p1 <- c(0.2, 0.5, 0.5)
get_details(design = design, n = 20, p1 = p1)
}
