% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_details.R
\name{get_details.exnex}
\alias{get_details.exnex}
\title{Get Details of a Basket Trial Simulation with the EXNEX Design}
\usage{
\method{get_details}{exnex}(design, n, p1, lambda, tau_scale, w, iter = 1000, data = NULL, ...)
}
\arguments{
\item{design}{An object of class \code{exnex}.}

\item{n}{The sample size per basket.}

\item{p1}{Probabilities under the alternative hypothesis.}

\item{lambda}{The posterior probability threshold.}

\item{tau_scale}{Standard deviation of the half normal prior exchangeability
distribution for the variance of the thetas.}

\item{w}{Fixed prior weight for the exchangeability part of the model.}

\item{iter}{The number of iterations in the simulation.}

\item{data}{An object of class \code{scenario_list} as returned by the
function \code{bhmbasket::simulateScenarios}.}

\item{...}{Further arguments.}
}
\value{
A list containing the rejection probabilites, posterior means,
mean squared errors and mean limits of HDI intervals for all baskets.
}
\description{
Get Details of a Basket Trial Simulation with the EXNEX Design
}
\examples{
design <- setup_exnex(k = 3, p0 = 0.2)
get_details(design = design, n = 20, p1 = c(0.2, 0.5, 0.5), lambda = 0.95,
  tau_scale = 1, w = 0.5)
}
