% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{setup_jsdglobal}
\alias{setup_jsdglobal}
\title{Setup Global JSD Design Object}
\usage{
setup_jsdglobal(k, p0, shape1 = 1, shape2 = 1)
}
\arguments{
\item{k}{The number of baskets.}

\item{p0}{A common probability under the null hypothesis.}

\item{shape1}{First common shape parameter of the beta prior.}

\item{shape2}{Second common shape parameter of the beta prior.}
}
\value{
An S3 object of class \code{jsdglobal}
}
\description{
Setup Global JSD Design Object
}
\examples{
design_jsdglobal <- setup_jsdglobal(k = 3, p0 = 0.2)
}
